Retrieve firmware version using float

+++ src/Core/Menu/Overlay.hpp
@@ -28,7 +28,7 @@ public:
    float m_FanSpeed{};
    memUsage_s m_MemoryUsage{};
    uint64_t m_KernelType{};
-   uint32_t m_FirmwareVersion{};
+   float m_FirmwareVersion{};
    uint16_t m_PayloadVersion{};
    static const uint64_t refreshDelay = 500;
